etcd_op/requirements_for_py3.txt
absl-py==1.0.0
etcd3==0.12.0
grpcio==1.42.0
prettytable==2.4.0
protobuf==3.19.1
six==1.16.0
tenacity==8.0.1
wcwidth==0.2.5
